refactor: extract LandingPage sub-components (F5)
Begin decomposing the 1096-line LandingPage: extract the self-contained
MarqueeTicker (live ticker strip) and the shared framer-motion animation
variants into frontend/src/components/landing/ (MarqueeTicker.js,
animations.js). LandingPage drops to 944 lines and imports them; the
remaining in-file sections use the imported variants. Pure code movement,
no behavior change.

Harden the characterization test with fake timers so the page's
timer-driven animations don't fire async state updates after the
synchronous assertions (removes intermittent flakiness). Full suite (82
tests) green across repeated runs; lint + build clean.
